+++ src/main/java/se/storkforge/petconnect/config/TwilioConfig.java
@@ -1,0 +1,28 @@
+package se.storkforge.petconnect.config;
+
+import com.twilio.Twilio;
+import jakarta.annotation.PostConstruct;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class TwilioConfig {
+
+    @Value("${twilio.account_sid}")
+    private String accountSid;
+
+    @Value("${twilio.auth_token}")
+    private String authToken;
+
+    @Value("${twilio.phone_number}")
+    private String fromPhoneNumber;
+
+    @PostConstruct
+    public void init() {
+        Twilio.init(accountSid, authToken);
+    }
+
+    public String getFromPhoneNumber() {
+        return fromPhoneNumber;
+    }
+}
