+++ src/main/java/se/storkforge/petconnect/service/SmsService.java
@@ -1,28 +1,16 @@
 package se.storkforge.petconnect.service;
 
-import com.twilio.Twilio;
-import jakarta.annotation.PostConstruct;
 import com.twilio.type.PhoneNumber;
 import com.twilio.rest.api.v2010.account.Message;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import se.storkforge.petconnect.config.TwilioConfig;
 
 @Service
 public class SmsService {
 
-    @Value("${twilio.account_sid}")
-    private String accountSid;
-
-    @Value("${twilio.auth_token}")
-    private String authToken;
-
-    @Value("${twilio.phone_number}")
-    private String fromPhoneNumber;
-
-    @PostConstruct
-    public void init() {
-        Twilio.init(accountSid, authToken);
-    }
+   @Autowired
+   private TwilioConfig twilioConfig;
 
     public void sendSms(String to, String messageText) {
         if (to == null || to.isEmpty()) {
@@ -30,9 +18,9 @@ public class SmsService {
         }
 
         try {
-            Message message = Message.creator(
+            Message.creator(
                     new PhoneNumber(to),
-                    new PhoneNumber(fromPhoneNumber),
+                    new PhoneNumber(twilioConfig.getFromPhoneNumber()),
                     messageText
 
             ).create();
