+++ src/main/resources/application.properties
@@ -49,4 +49,4 @@ spring.flyway.target=latest
 # === Twilio
 twilio.account_sid=your_account_sid
 twilio.auth_token=your_auth_token
-twilio.phone_number=+1234567890
+twilio.phone_number=your_twilio_phone_number
